use penumbra protos for Address
Pulls in penumbra protos from BSR to use the Address struct to wrangle
bytes and strings. Turns out it's pretty simple: we can be super naive
about the bytes/string conversion, disregarding nuances of bech32 vs
bech32m. We just need to pass the raw bytes of the string to the proto
struct, and we're good to go. Let's see if we can't update the
interchaintest code to do the same.