demo pclientd integration
A bit slipshop, in that it hardcodes a local (to my machine) penumbra
address, and requires running pclientd out of band. Still, it's a useful
demonstration of the bytes -> string conversion via bech32m. As for
automating the pclientd integration, that's precisely what
interchaintest is for, and this demo should support further progress on
that front.