demo golang str -> bytes -> str round-trip
Adds a working example of taking a Penumbra address as string, decoding
it to bytes, and re-encoding as the original string, all via the bech32m
library. This is helpful for demonstration purposes, but notably the
intermediate bytes representation is *not* suitable for passing to the
gRPC methods. Still need a method for doing that.